computer-use/requirements.txt
openai>=1.68.2
pyautogui>=0.9.54
Pillow>=11.1.0
paramiko==3.5.1
dotenv==0.9.9